add golang tests for the sqlite database
This tests the numbers of books, chapters, and verses in each
translation. It makes invalid assertions as each translation may have a
different number of verses in each chater as compared to the t_kjv.
However I believe we should acheive consistency with the t_kjv verses
for ease of use.