Fix book_info.sql and dump entire database to sql and sqlite3
book_info.sql had some sqlite3 isms, that was preventing mysql from
loading it. I noticed this when using the new docker-compose file. After
fixing this, I did a complete to to bible-mysql.sql so that it matches
the sum of what is in sql/.

At the same time, I updated the README to reflect that sqlite is not a
simplified database, it is really just a different format. It contains
all the information from the mysql database, with the addition of the
cross reference (The readme did say cross-reference was included, but it
actually wasnt included in reality. I have now included it)